feat: C-c C-c to close window via hook
You can now close *marginalia* buffer with C-c C-c.  This is achieved by
adding `kill-buffer-and-window` to `org-ctrl-c-ctrl-c-final-hook`.  As the
functions in this hook are called after all the other default ones, C-c C-c to
close respects the normal org operations; it needs to be done in a context
where none of the built-in ones are effective.

Update: marginalia (README)